refactor(db): move getConn/runQuery to lib/db/primitives; break connectionmanager->components cycle
