+++ frontend/.cycle-baseline
@@ -1,1 +1,1 @@
-34
+33
