+++ frontend/app/api/export/ctfs-sql/[schema]/[plotID]/[censusID]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { HTTPResponses } from '@/config/macros';
 import { isValidSchema, safeFormatQuery } from '@/config/utils/sqlsecurity';
-import { getConn } from '@/components/processors/processormacros';
+import { getConn } from '@/lib/db/primitives';
 import { auth } from '@/auth';
 import { requireSession, getSessionUserId } from '@/lib/auth-helpers';
 import ailogger from '@/ailogger';
