+++ frontend/config/orphaneddatacleanup.ts
@@ -11,7 +11,7 @@
  */
 
 import { PoolConnection } from 'mysql2/promise';
-import { getConn, runQuery } from '@/components/processors/processormacros';
+import { getConn, runQuery } from '@/lib/db/primitives';
 import ailogger from '@/ailogger';
 
 export interface OrphanedDataStats {
